feat: remove Telegram-specific UI — Stories, Stars, Gifts, Ads, Boosts, Monetization, Paid media
Remove 7 categories of Telegram-specific features per TZ §20.2:
- Stories (viewer, ribbon, creation, statistics, stealth mode)
- Stars/TON (wallet, payments, subscriptions, transactions)
- Gifts/NFT/Collectibles (shop, auction, craft, transfer, resale)
- Ads/Sponsored messages (sponsored messages, bot ads, report/about modals)
- Boost/Giveaway (boost modal, statistics, giveaway creation)
- Channel Monetization (revenue stats, withdrawal, verification)
- Paid reactions/media (paid reactions, paid media overlay, price confirm)

Also fixes:
- Voice message waveform: backend now returns number[] instead of base64
- Audio.tsx: skip redundant decodeWaveform for Saturn data
- fetchExportedChatInvites: fix undefined chatId (peer.id extraction)
- Search: guard against empty query strings

386 files changed, ~42800 lines removed, 0 new TypeScript errors.

+++ services/media/internal/model/models.go
@@ -86,7 +86,7 @@ type MediaResponse struct {
 	Width            *int     `json:"width,omitempty"`
 	Height           *int     `json:"height,omitempty"`
 	DurationSeconds  *float64 `json:"duration_seconds,omitempty"`
-	WaveformData     []byte   `json:"waveform_data,omitempty"`
+	WaveformData     []int    `json:"waveform_data,omitempty"`
 	ProcessingStatus string   `json:"processing_status"`
 }
